Move TextGeometry, FontLoader and Font to examples. (#22560)
