+++ docs/api/en/extras/core/Interpolations.html
@@ -28,7 +28,7 @@
 		t -- interpolation weight.<br />
 		p0, p1, p2 -- the starting, control and end points defining the curve.<br /><br />
 
-		Used internally by [page:QuadraticBezierCurve3 QuadraticBezierCurve3], [page:QuadraticBezierCurve QuadraticBezierCurve] and [page:Font Font].
+		Used internally by [page:QuadraticBezierCurve3 QuadraticBezierCurve3] and [page:QuadraticBezierCurve QuadraticBezierCurve].
 		</p>
 
 		<h3>[method:Float CubicBezier]( [param:Float t], [param:Float p0], [param:Float p1], [param:Float p2], [param:Float p3] )</h3>
@@ -36,7 +36,7 @@
 		t -- interpolation weight.<br />
 		p0, p1, p2, p3 -- the starting, control(twice) and end points defining the curve.<br /><br />
 
-		Used internally by [page:CubicBezierCurve3 CubicBezierCurve3], [page:CubicBezierCurve CubicBezierCurve] and [page:Font Font].
+		Used internally by [page:CubicBezierCurve3 CubicBezierCurve3] and [page:CubicBezierCurve CubicBezierCurve].
 		</p>
 
 		<h2>Source</h2>
