+++ docs/api/en/extras/core/ShapePath.html
@@ -11,8 +11,7 @@
 
 		<p class="desc">
 		This class is used to convert a series of shapes to an array of [page:Path]s, for example an SVG shape to a
-		path (see the example below). It is used internally by [page:Font] to convert a font in JSON format to a
-		series of paths.
+		path (see the example below).
 		</p>
 
 		<h2>Examples</h2>
