+++ docs/api/ko/extras/core/Interpolations.html
@@ -28,7 +28,7 @@
 		t -- 보간 가중치.<br />
 		p0, p1, p2 -- 곡선을 정의하는 시작, 조절, 끝 점입니다.<br /><br />
 
-		[page:QuadraticBezierCurve3 QuadraticBezierCurve3], [page:QuadraticBezierCurve QuadraticBezierCurve] 및 [page:Font Font]에서 내부적으로 사용됩니다.
+		[page:QuadraticBezierCurve3 QuadraticBezierCurve3] 및 [page:QuadraticBezierCurve QuadraticBezierCurve]에서 내부적으로 사용됩니다.
 		</p>
 
 		<h3>[method:Float CubicBezier]( [param:Float t], [param:Float p0], [param:Float p1], [param:Float p2], [param:Float p3] )</h3>
@@ -36,7 +36,7 @@
 		t -- 보간 가중치.<br />
 		p0, p1, p2, p3 -- 곡선을 정의하는 시작, 조절(두 번), 끝 점입니다.<br /><br />
 
-		[page:CubicBezierCurve3 CubicBezierCurve3], [page:CubicBezierCurve CubicBezierCurve] 및 [page:Font Font]에서 내부적으로 사용됩니다.
+		[page:CubicBezierCurve3 CubicBezierCurve3] 및 [page:CubicBezierCurve CubicBezierCurve]에서 내부적으로 사용됩니다.
 		</p>
 
 		<h2>소스코드</h2>
