--- docs/api/zh/extras/core/Font.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh">
-	<head>
-		<meta charset="utf-8" />
-		<base href="../../../../" />
-		<script src="page.js"></script>
-		<link type="text/css" rel="stylesheet" href="page.css" />
-	</head>
-	<body>
-		<h1>字体（[name]）</h1>
-
-		<p class="desc">
-		以JSON格式，创建一系列的[page:Shape Shape]（形状）来表示一个字体。
-		<br /><br />
-
-		该类在内部由[page:FontLoader]所使用。
-		</p>
-
-		<h2>例子</h2>
-
-		<p>
-		[example:webgl_geometry_text_shapes geometry / text / shapes ]
-		</p>
-
-		<h2>构造函数</h2>
-
-		<h3>[name]( data )</h3>
-		<p>
-		data -- 表示字体的JSON数据。<br /><br />
-
-		这一构造函数创建一个新的[name]，它是一个[page:Shape Shapes]数组。
-		</p>
-
-		<h2>属性</h2>
-
-		<h3>[property:Array data]</h3>
-		<p>传入到构造函数的JSON数据。</p>
-
-		<h2>方法</h2>
-
-		<h3>[method:null generateShapes]( [param:String text], [param:Float size] )</h3>
-		<p>
-			[page:String text] -- 文本字符串。<br />
-			[page:Float size] -- （可选）[page:Shape Shapes]的缩放，默认值为*100*。<br />
-
-			创建一个[page:Shape Shapes]数组，表示使用字体的文本。
-		</p>
-
-		<h2>源代码</h2>
-
-		<p>
-			[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
-		</p>
-	</body>
-</html>
