+++ docs/api/zh/extras/core/Interpolations.html
@@ -28,7 +28,7 @@
 		t -- 插值权重<br />
 		p0, p1, p2 -- 定义了该曲线的起始点、控制点和终止点。<br /><br />
 
-		在内部由[page:QuadraticBezierCurve3 QuadraticBezierCurve3]、[page:QuadraticBezierCurve QuadraticBezierCurve]和[page:Font Font]所使用。
+		在内部由[page:QuadraticBezierCurve3 QuadraticBezierCurve3]和[page:QuadraticBezierCurve QuadraticBezierCurve]所使用。
 		</p>
 
 		<h3>[method:Float CubicBezier]( [param:Float t], [param:Float p0], [param:Float p1], [param:Float p2], [param:Float p3] )</h3>
@@ -36,7 +36,7 @@
 		t -- 插值权重<br />
 		p0, p1, p2, p3 -- 定义了该曲线的起始点、两个控制点和终止点。<br /><br />
 
-		在内部由[page:CubicBezierCurve3 CubicBezierCurve3]、[page:CubicBezierCurve CubicBezierCurve]和[page:Font Font]所使用。
+		在内部由[page:CubicBezierCurve3 CubicBezierCurve3]和[page:CubicBezierCurve CubicBezierCurve]所使用。
 		</p>
 
 		<h2>源代码</h2>
