+++ docs/api/zh/extras/core/ShapePath.html
@@ -11,7 +11,6 @@
 
 		<p class="desc">
 		该类用于转换一系列的形状为一个[page:Path]数组，例如转换SVG中的Path为three.js中的Path（请参阅下方的example）。
-		在内部它由[page:Font]所使用，用于将JSON字体转换为一系列路径。
 		</p>
 
 		<h2>例子</h2>
