+++ docs/examples/en/loaders/FontLoader.html
@@ -12,7 +12,7 @@
 		<h1>[name]</h1>
 
 		<p class="desc">
-		Class for loading a font in JSON format. Returns a [page:Font Font], which is an
+		Class for loading a font in JSON format. Returns a font, which is an
 		array of [page:Shape Shapes] representing the font.
 		This uses the [page:FileLoader] internally for loading files. <br /><br />
 
@@ -22,7 +22,7 @@
 		<h2>Code Example</h2>
 
 		<code>
-		const loader = new THREE.FontLoader();
+		const loader = new FontLoader();
 		const font = loader.load(
 			// resource URL
 			'fonts/helvetiker_bold.typeface.json',
@@ -30,7 +30,7 @@
 			// onLoad callback
 			function ( font ) {
 				// do something with the font
-				scene.add( font );
+				console.log( font );
 			},
 
 			// onProgress callback
@@ -70,7 +70,7 @@
 		<p>
 		[page:String url] — the path or URL to the file. This can also be a
 			[link:https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/Data_URIs Data URI].<br />
-		[page:Function onLoad] — Will be called when load completes. The argument will be the loaded [page:Font font].<br />
+		[page:Function onLoad] — Will be called when load completes. The argument will be the loaded font.<br />
 		[page:Function onProgress] — Will be called while load progresses. The argument will be the XMLHttpRequest instance, which contains .[page:Integer total] and .[page:Integer loaded] bytes.<br />
 		[page:Function onError] — Will be called when load errors.<br /><br />
 
@@ -80,13 +80,13 @@
 		<h3>[method:Font parse]( [param:Object json] )</h3>
 		<p>
 		[page:Object json] — The <em>JSON</em> structure to parse.<br /><br />
-		Parse a <em>JSON</em> structure and return a [page:Font].
+		Parse a <em>JSON</em> structure and return a font.
 		</p>
 
 		<h2>Source</h2>
 
 		<p>
-			[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+			[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/FontLoader.js examples/jsm/loaders/FontLoader.js]
 		</p>
 	</body>
 </html>
