+++ docs/scenes/geometry-browser.html
@@ -38,7 +38,6 @@
 				DoubleSide,
 				ExtrudeGeometry,
 				Float32BufferAttribute,
-				FontLoader,
 				Group,
 				IcosahedronGeometry,
 				LatheGeometry,
@@ -56,7 +55,6 @@
 				ShapeGeometry,
 				SphereGeometry,
 				TetrahedronGeometry,
-				TextGeometry,
 				TorusGeometry,
 				TorusKnotGeometry,
 				TubeGeometry,
@@ -68,7 +66,6 @@
 
 			import { GUI } from '../../examples/jsm/libs/dat.gui.module.js';
 			import { OrbitControls } from '../../examples/jsm/controls/OrbitControls.js';
-			import { ParametricGeometries } from '../../examples/jsm/geometries/ParametricGeometries.js';
 
 			const twoPi = Math.PI * 2;
 
@@ -510,78 +507,6 @@
 
 					folder.add( data, 'radius', 1, 20 ).onChange( generateGeometry );
 					folder.add( data, 'detail', 0, 5 ).step( 1 ).onChange( generateGeometry );
-
-					generateGeometry();
-
-				},
-
-				TextGeometry: function ( mesh ) {
-
-					const data = {
-						text: 'TextGeometry',
-						size: 5,
-						height: 2,
-						curveSegments: 12,
-						font: 'helvetiker',
-						weight: 'regular',
-						bevelEnabled: false,
-						bevelThickness: 1,
-						bevelSize: 0.5,
-						bevelOffset: 0.0,
-						bevelSegments: 3
-					};
-
-					const fonts = [
-						'helvetiker',
-						'optimer',
-						'gentilis',
-						'droid/droid_serif'
-					];
-
-					const weights = [
-						'regular', 'bold'
-					];
-
-					function generateGeometry() {
-
-						const loader = new FontLoader();
-						loader.load( '../../examples/fonts/' + data.font + '_' + data.weight + '.typeface.json', function ( font ) {
-
-							const geometry = new TextGeometry( data.text, {
-								font: font,
-								size: data.size,
-								height: data.height,
-								curveSegments: data.curveSegments,
-								bevelEnabled: data.bevelEnabled,
-								bevelThickness: data.bevelThickness,
-								bevelSize: data.bevelSize,
-								bevelOffset: data.bevelOffset,
-								bevelSegments: data.bevelSegments
-							} );
-							geometry.center();
-
-							updateGroupGeometry( mesh, geometry );
-
-						} );
-
-					}
-
-					//Hide the wireframe
-					mesh.children[ 0 ].visible = false;
-
-					const folder = gui.addFolder( 'THREE.TextGeometry' );
-
-					folder.add( data, 'text' ).onChange( generateGeometry );
-					folder.add( data, 'size', 1, 30 ).onChange( generateGeometry );
-					folder.add( data, 'height', 1, 20 ).onChange( generateGeometry );
-					folder.add( data, 'curveSegments', 1, 20 ).step( 1 ).onChange( generateGeometry );
-					folder.add( data, 'font', fonts ).onChange( generateGeometry );
-					folder.add( data, 'weight', weights ).onChange( generateGeometry );
-					folder.add( data, 'bevelEnabled' ).onChange( generateGeometry );
-					folder.add( data, 'bevelThickness', 0.1, 3 ).onChange( generateGeometry );
-					folder.add( data, 'bevelSize', 0, 3 ).onChange( generateGeometry );
-					folder.add( data, 'bevelOffset', - 0.5, 1.5 ).onChange( generateGeometry );
-					folder.add( data, 'bevelSegments', 0, 8 ).step( 1 ).onChange( generateGeometry );
 
 					generateGeometry();
