+++ examples/webgl_custom_attributes_lines.html
@@ -51,13 +51,16 @@
 
 			import * as THREE from '../build/three.module.js';
 
+			import { FontLoader } from './jsm/loaders/FontLoader.js';
+			import { TextGeometry } from './jsm/geometries/TextGeometry.js';
+
 			import Stats from './jsm/libs/stats.module.js';
 
 			let renderer, scene, camera, stats;
 
 			let line, uniforms;
 
-			const loader = new THREE.FontLoader();
+			const loader = new FontLoader();
 			loader.load( 'fonts/helvetiker_bold.typeface.json', function ( font ) {
 
 				init( font );
@@ -93,7 +96,7 @@
 				} );
 
 
-				const geometry = new THREE.TextGeometry( 'three.js', {
+				const geometry = new TextGeometry( 'three.js', {
 
 					font: font,
