+++ examples/webgl_geometry_text_shapes.html
@@ -25,6 +25,7 @@
 			import * as THREE from '../build/three.module.js';
 
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
+			import { FontLoader } from './jsm/loaders/FontLoader.js';
 
 			let camera, scene, renderer;
 
@@ -39,7 +40,7 @@
 				scene = new THREE.Scene();
 				scene.background = new THREE.Color( 0xf0f0f0 );
 
-				const loader = new THREE.FontLoader();
+				const loader = new FontLoader();
 				loader.load( 'fonts/helvetiker_regular.typeface.json', function ( font ) {
 
 					const color = 0x006699;
