+++ examples/webgl_shadowmap_performance.html
@@ -22,6 +22,8 @@
 
 			import { FirstPersonControls } from './jsm/controls/FirstPersonControls.js';
 			import { GLTFLoader } from './jsm/loaders/GLTFLoader.js';
+			import { FontLoader } from './jsm/loaders/FontLoader.js';
+			import { TextGeometry } from './jsm/geometries/TextGeometry.js';
 
 			const SHADOW_MAP_WIDTH = 2048, SHADOW_MAP_HEIGHT = 1024;
 
@@ -154,10 +156,10 @@
 
 				// TEXT
 
-				const loader = new THREE.FontLoader();
+				const loader = new FontLoader();
 				loader.load( 'fonts/helvetiker_bold.typeface.json', function ( font ) {
 
-					const textGeo = new THREE.TextGeometry( "THREE.JS", {
+					const textGeo = new TextGeometry( "THREE.JS", {
 
 						font: font,
