--- test/unit/src/extras/core/Font.tests.js
@@ -1,35 +1,0 @@
-/* global QUnit */
-
-import { Font } from '../../../../../src/extras/core/Font';
-
-export default QUnit.module( 'Extras', () => {
-
-	QUnit.module( 'Core', () => {
-
-		QUnit.module( 'Font', () => {
-
-			// INSTANCING
-			QUnit.todo( "Instancing", ( assert ) => {
-
-				assert.ok( false, "everything's gonna be alright" );
-
-			} );
-
-			// PUBLIC STUFF
-			QUnit.todo( "isFont", ( assert ) => {
-
-				assert.ok( false, "everything's gonna be alright" );
-
-			} );
-
-			QUnit.todo( "generateShapes", ( assert ) => {
-
-				assert.ok( false, "everything's gonna be alright" );
-
-			} );
-
-		} );
-
-	} );
-
-} );
