--- test/unit/src/loaders/FontLoader.tests.js
@@ -1,37 +1,0 @@
-/* global QUnit */
-
-import { FontLoader } from '../../../../src/loaders/FontLoader';
-
-export default QUnit.module( 'Loaders', () => {
-
-	QUnit.module( 'FontLoader', () => {
-
-		// INSTANCING
-		QUnit.todo( "Instancing", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		// PUBLIC STUFF
-		QUnit.todo( "load", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		QUnit.todo( "parse", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-		QUnit.todo( "setPath", ( assert ) => {
-
-			assert.ok( false, "everything's gonna be alright" );
-
-		} );
-
-	} );
-
-} );
