+++ test/unit/three.source.unit.js
@@ -64,7 +64,6 @@ import './src/extras/ShapeUtils.tests';
 //src/extras/core
 import './src/extras/core/Curve.tests';
 import './src/extras/core/CurvePath.tests';
-import './src/extras/core/Font.tests';
 import './src/extras/core/Interpolations.tests';
 import './src/extras/core/Path.tests';
 import './src/extras/core/Shape.tests';
@@ -103,7 +102,6 @@ import './src/geometries/RingGeometry.tests';
 import './src/geometries/ShapeGeometry.tests';
 import './src/geometries/SphereGeometry.tests';
 import './src/geometries/TetrahedronGeometry.tests';
-import './src/geometries/TextGeometry.tests';
 import './src/geometries/TorusGeometry.tests';
 import './src/geometries/TorusKnotGeometry.tests';
 import './src/geometries/TubeGeometry.tests';
@@ -148,7 +146,6 @@ import './src/loaders/CompressedTextureLoader.tests';
 import './src/loaders/CubeTextureLoader.tests';
 import './src/loaders/DataTextureLoader.tests';
 import './src/loaders/FileLoader.tests';
-import './src/loaders/FontLoader.tests';
 import './src/loaders/ImageLoader.tests';
 import './src/loaders/Loader.tests';
 import './src/loaders/LoaderUtils.tests';
